+++ eras/dijkstra/impl/src/Cardano/Ledger/Dijkstra.hs
@@ -18,6 +18,7 @@ import Cardano.Ledger.Block (EraBlockHeader)
 import Cardano.Ledger.Conway.Governance (RunConwayRatify)
 import Cardano.Ledger.Dijkstra.BlockBody ()
 import Cardano.Ledger.Dijkstra.Era
+import Cardano.Ledger.Dijkstra.Forecast ()
 import Cardano.Ledger.Dijkstra.Genesis ()
 import Cardano.Ledger.Dijkstra.Governance ()
 import Cardano.Ledger.Dijkstra.Rules (