+++ eras/babbage/impl/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## 1.13.1.0
 
+* Deprecate the use of `GetLedgerView`:
+  - Add `BabbageForecast` to deprecate `LedgerView` for Praos.
+    + `mkBabbageForecast`
+    + `bfPoolDistrL`
+    + `bfMaxBlockHeaderSizeL`
+    + `bfMaxBlockBodySizeL`
+    + `bfProtocolVersionL`
+  - Add `EraForecast` instance with `BabbageForecast`.
 * Add `validateScriptsWellFormedTxOuts`
 
 ## 1.13.0.0